requirements.txt
Flask
Flask-CORS
numpy
Pillow
pydicom
tensorflow
opencv-python
